Updated to use floats

+++ BubblesDivePlanner/ViewModels/Model/Plan/DiveModels/IDiveModel.cs
@@ -4,12 +4,12 @@ namespace BubblesDivePlanner.ViewModels.Model.Plan.DiveModels
     {
         string Name { get; }
         byte Compartments { get; }
-        double[] NitrogenHalfTimes { get; }
-        double[] HeliumHalfTimes { get; }
-        double[] AValuesNitrogen { get; }
-        double[] BValuesNitrogen { get; }
-        double[] AValuesHelium { get; }
-        double[] BValuesHelium { get; }
+        float[] NitrogenHalfTimes { get; }
+        float[] HeliumHalfTimes { get; }
+        float[] AValuesNitrogen { get; }
+        float[] BValuesNitrogen { get; }
+        float[] AValuesHelium { get; }
+        float[] BValuesHelium { get; }
         IDiveProfile DiveProfile { get; }
     }
 }
